Workaround 'casting unsigned e->e_phnum to wider int type' code defect FP
* dyn_load.c [DYNAMIC_LOADING && !ANY_MSWIN && (DGUX || HURD || NACL
|| SCO_ELF || SERENITY || (ANY_BSD || LINUX) && __ELF__)
&& !USE_PROC_FOR_LIBRARIES && HAVE_DL_ITERATE_PHDR]
(GC_register_dynamic_libraries): Change type of `i` local variable from
`int` to `size_t`.

+++ dyn_load.c
@@ -874,7 +874,7 @@ GC_register_dynamic_libraries(void)
     const ElfW(Ehdr) * e;
     const ElfW(Phdr) * p;
     ptr_t load_ptr = (ptr_t)lm->l_addr;
-    int i;
+    size_t i;
 
 #    ifdef HOST_ANDROID
     if (NULL == load_ptr)
@@ -882,7 +882,7 @@ GC_register_dynamic_libraries(void)
 #    endif
     e = (ElfW(Ehdr) *)load_ptr;
     p = (ElfW(Phdr) *)(load_ptr + e->e_phoff);
-    for (i = 0; i < (int)e->e_phnum; i++, p++) {
+    for (i = 0; i < (size_t)e->e_phnum; i++, p++) {
       ptr_t start;
 
       switch (p->p_type) {
